change minor shortcomings
